refact: Move logic for running multiple speedup to a separate module
